refactor: split ui panel modules
