+++ src/ui/panels/mod.rs
@@ -1,0 +1,11 @@
+//! Panel rendering helpers.
+
+mod commands;
+mod details;
+mod device_lists;
+mod logs;
+
+pub(crate) use commands::{render_device_commands, render_log_commands};
+pub(crate) use details::render_device_details_panel;
+pub(crate) use device_lists::{render_android_panel, render_ios_panel};
+pub(crate) use logs::render_log_panel;
